feat: :fire: Took a bit, but finally hit it.
Long story short is..... This is a Python class that implements a method to find the median of two sorted arrays nums1 and nums2. It works by first finding the shortest array and setting up some variables to perform a binary search. It then performs a binary search on the shorter array to find the median point of both arrays. Based on that, it determines the median value, which is returned in float format.